refactor(rag): extract context builder into rag/context.rs
Move build_context into context.rs; public path application::rag::build_context
preserved via re-export. mod.rs now holds only types + query/run_action
orchestration. Behavior unchanged.
